Move rpc pending data wrappers to own file
